+++ src/pages/Reservation/ReservationListPage.jsx
@@ -2,7 +2,7 @@ import BackButton from "../../components/Button/BackButton";
 import PageContainer from "../../components/Common/PageContainer";
 import ReservationList from "../../components/Reservation/ReservationList.jsx";
 
-const UntactReservePage = () => {
+const ReservationListPage = () => {
   return (
     <PageContainer>
       <BackButton text="대시보드로 돌아가기" />
@@ -11,4 +11,4 @@ const UntactReservePage = () => {
   );
 };
 
-export default UntactReservePage;
+export default ReservationListPage;
